requirements.txt
hikari
hikari-lightbulb
lavasnek-rs
uvloop
python-dotenv
psutil
mcipc
